+++ LookupAnything/Framework/Data/PuzzleSolutionsData.cs
@@ -1,0 +1,11 @@
+using StardewValley.Locations;
+
+namespace Pathoschild.Stardew.LookupAnything.Framework.Data
+{
+    /// <summary>Solutions for hardcoded puzzles.</summary>
+    internal class PuzzleSolutionsData
+    {
+        /// <summary>The sequence of flute block pitches for the <see cref="IslandSouthEast"/> mermaid music puzzle</summary>
+        public int[] IslandMermaidFluteBlockSequence { get; set; }
+    }
+}